Fix Winx86 compilers fail to build unit test: linker "file too big" errors. Fix is to *move* the tests out of line.
